mifica-streamlit/requirements.txt
streamlit
requests
PyJWT
Pillow
plotly
pandas
numpy
